+++ java-manta-it/src/test/java/com/joyent/manta/client/MantaClientIT.java
@@ -230,6 +230,33 @@ public class MantaClientIT {
     }
 
     @Test
+    public final void canGetZeroByteFileAsInputStream() throws IOException {
+        final String name = UUID.randomUUID().toString();
+        final String path = testPathPrefix + name;
+        final byte[] empty = new byte[0];
+
+        mantaClient.put(path, empty);
+        try (final MantaObjectInputStream gotObject = mantaClient.getAsInputStream(path)) {
+            Assert.assertEquals(gotObject.getContentLength().longValue(), 0);
+            byte[] actualBytes = IOUtils.toByteArray(gotObject);
+            Assert.assertEquals(actualBytes, empty,
+                    "Actual object was not the expected 0 bytes");
+        }
+    }
+
+    @Test
+    public final void canGetZeroByteFileAsString() throws IOException {
+        final String name = UUID.randomUUID().toString();
+        final String path = testPathPrefix + name;
+        final byte[] empty = new byte[0];
+
+        mantaClient.put(path, empty);
+        final String actual = mantaClient.getAsString(path, StandardCharsets.UTF_8);
+        Assert.assertEquals(actual, "",
+                "Empty string not returned as expected");
+    }
+
+    @Test
     public final void testContentTypeSetByFilename() throws IOException {
         final String name = UUID.randomUUID().toString() + ".html";
         final String path = testPathPrefix + name;